[change] add BrouserRouter in App.tsx
