+++ src/components/Setsu/index.tsx
@@ -1,0 +1,10 @@
+import React, { FC, useState } from 'react';
+import { Button } from 'semantic-ui-react';
+
+const Setsu: FC<{}> = () => (
+  <>
+    <Button>DO IT</Button>
+  </>
+);
+
+export default Setsu;
